FW-003 Notification fix Overviewfrag [LH,GF]
